+++ Assets/Scripts/Intelligence/Actions/BougerRandomScout.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 public class BougerRandomScout : BougerRandom
 {
 
-    public BougerRandomScout(long p_duree, Quaternion p_direction) : base(p_duree, p_direction) {
+    public BougerRandomScout(float p_duree, Quaternion p_direction) : base(p_duree, p_direction) {
     }
 
     public static BougerRandomScout createRandom()
